+++ storage/innobase/mem/mem0mem.cc
@@ -126,7 +126,7 @@ mem_heap_printf_low(
 
 				val = va_arg(ap, unsigned long);
 
-				plen = sprintf(tmp, "%lu", val);
+				plen = size_t(sprintf(tmp, "%lu", val));
 				len += plen;
 
 				if (buf) {
